separated audiotypes and audioenginetypes
