+++ framework/audio/engine/audiotypes.h
@@ -19,14 +19,19 @@
 #ifndef MU_AUDIO_AUDIOTYPES_H
 #define MU_AUDIO_AUDIOTYPES_H
 
-#include <memory>
-#include <map>
+#include <string>
 
-#include "realfn.h"
+#include "internal/audioenginetypes.h"
 
 namespace mu {
 namespace audio {
-namespace engine {
+enum class PlayStatus {
+    UNDEFINED = 0,
+    STOPED,
+    PLAYING,
+    PAUSED
+};
+
 struct LoopRegion
 {
     float begin = 0.0f;
@@ -35,87 +40,6 @@ struct LoopRegion
     bool isValid() const { return !(begin < 0.f) && end > begin; }
 };
 
-// === Args ===
-class Args
-{
-public:
-
-    template<typename T>
-    static Args make_arg1(const T& val)
-    {
-        Args d;
-        d.setArg<T>(0, val);
-        return d;
-    }
-
-    template<typename T1, typename T2>
-    static Args make_arg2(const T1& val1, const T2& val2)
-    {
-        Args d;
-        d.setArg<T1>(0, val1);
-        d.setArg<T2>(1, val2);
-        return d;
-    }
-
-    template<typename T>
-    void setArg(int i, const T& val)
-    {
-        IArg* p = new Arg<T>(val);
-        m_args[i] = std::shared_ptr<IArg>(p);
-    }
-
-    template<typename T>
-    T arg(int i = 0, T def = T()) const
-    {
-        auto it = m_args.find(i);
-        if (it == m_args.cend()) {
-            return def;
-        }
-        IArg* p = it->second.get();
-        Arg<T>* d = reinterpret_cast<Arg<T>*>(p);
-        return d->val;
-    }
-
-    bool hasArg(int i) const
-    {
-        auto it = m_args.find(i);
-        if (it != m_args.cend()) {
-            return true;
-        }
-        return false;
-    }
-
-    int count() const
-    {
-        return int(m_args.size());
-    }
-
-    void clear()
-    {
-        m_args.clear();
-    }
-
-    void swap(Args& other)
-    {
-        m_args.swap(other.m_args);
-    }
-
-    struct IArg {
-        virtual ~IArg() = default;
-    };
-
-    template<typename T>
-    struct Arg : public IArg {
-        T val;
-        Arg(const T& v)
-            : IArg(), val(v) {}
-    };
-
-private:
-    std::map<int, std::shared_ptr<IArg> > m_args;
-};
-
-// === Context ===
 enum class CtxKey {
     Silent = 0,
     HasEnded = 1,
@@ -128,7 +52,7 @@ enum class CtxKey {
 
 struct Context
 {
-    Args args;
+    engine::Args args;
 
     template<typename T>
     void set(CtxKey k, const T& t)
@@ -179,7 +103,6 @@ struct Context
         return str;
     }
 };
-}
 }
 }
